+++ cpp/src/link_analysis/pagerank.cu
@@ -70,7 +70,7 @@ bool pagerankIteration(IndexType n, IndexType e, IndexType *cscPtr, IndexType *c
 }
 
 template <typename IndexType, typename ValueType>
-int pagerank(IndexType n, IndexType e, IndexType *cscPtr, IndexType *cscInd, ValueType *cscVal,
+int pagerankSolver(IndexType n, IndexType e, IndexType *cscPtr, IndexType *cscInd, ValueType *cscVal,
              IndexType *prsVtx, ValueType *prsVal, IndexType prsLen, bool has_personalization,
              ValueType alpha, ValueType *a, bool has_guess, float tolerance, int max_iter,
              ValueType * &pagerank_vector, ValueType * &residual) {
@@ -175,18 +175,16 @@ int pagerank(IndexType n, IndexType e, IndexType *cscPtr, IndexType *cscInd, Val
   return converged ? 0 : 1;
 }
 
-//template int pagerank<int, half> (  int n, int e, int *cscPtr, int *cscInd,half *cscVal, half alpha, half *a, bool has_guess, float tolerance, int max_iter, half * &pagerank_vector, half * &residual);
-template int pagerank<int, float> (  int n, int e, int *cscPtr, int *cscInd,float *cscVal,
+//template int pagerankSolver<int, half> (  int n, int e, int *cscPtr, int *cscInd,half *cscVal, half alpha, half *a, bool has_guess, float tolerance, int max_iter, half * &pagerank_vector, half * &residual);
+template int pagerankSolver<int, float> (  int n, int e, int *cscPtr, int *cscInd,float *cscVal,
         int *prsVtx, float *prsVal, int prsLen, bool has_personalization,
         float alpha, float *a, bool has_guess, float tolerance, int max_iter, float * &pagerank_vector, float * &residual);
-template int pagerank<int, double> (  int n, int e, int *cscPtr, int *cscInd,double *cscVal,
+template int pagerankSolver<int, double> (  int n, int e, int *cscPtr, int *cscInd,double *cscVal,
         int *prsVtx,  double *prsVal, int prsLen, bool has_personalization,
         double alpha, double *a, bool has_guess, float tolerance, int max_iter, double * &pagerank_vector, double * &residual);
 
-} } //namespace
-
 template <typename WT>
-gdf_error gdf_pagerank_impl (gdf_graph *graph,
+void pagerank_impl (Graph *graph,
                       gdf_column *pagerank,
                       gdf_column *personalization_subset, gdf_column *personalization_values,
                       float alpha = 0.85,
@@ -220,7 +218,7 @@ gdf_error gdf_pagerank_impl (gdf_graph *graph,
   WT *residual = &res;
 
   if (graph->transposedAdjList == nullptr) {
-    gdf_add_transposed_adj_list(graph);
+    add_transposed_adj_list(graph);
   }
   cudaStream_t stream{nullptr};
   ALLOC_TRY((void**)&d_leaf_vector, sizeof(WT) * m, stream);
@@ -232,26 +230,26 @@ gdf_error gdf_pagerank_impl (gdf_graph *graph,
 #endif
 
   //  The templating for HT_matrix_csc_coo assumes that m, nnz and data are all the same type
-  cugraph::detail::HT_matrix_csc_coo(m, nnz, (int *)graph->transposedAdjList->offsets->data, (int *)graph->transposedAdjList->indices->data, d_val, d_leaf_vector);
+  HT_matrix_csc_coo(m, nnz, (int *)graph->transposedAdjList->offsets->data, (int *)graph->transposedAdjList->indices->data, d_val, d_leaf_vector);
 
   if (has_guess)
   {
     CUGRAPH_EXPECTS( pagerank->data != nullptr, "Column must be valid" );
-    cugraph::detail::copy<WT>(m, (WT*)pagerank->data, d_pr);
+    copy<WT>(m, (WT*)pagerank->data, d_pr);
   }
 
-  status = cugraph::detail::pagerank<int32_t,WT>( m,nnz, (int*)graph->transposedAdjList->offsets->data, (int*)graph->transposedAdjList->indices->data, d_val,
+  status = pagerankSolver<int32_t,WT>( m,nnz, (int*)graph->transposedAdjList->offsets->data, (int*)graph->transposedAdjList->indices->data, d_val,
           prsVtx, prsVal, prsLen, has_personalization,
     alpha, d_leaf_vector, has_guess, tolerance, max_iter, d_pr, residual);
 
   if (status !=0)
     switch ( status ) {
-      case -1: std::cerr<< "Error : bad parameters in Pagerank"<<std::endl; return GDF_CUDA_ERROR;
-      case 1: std::cerr<< "Warning : Pagerank did not reached the desired tolerance"<<std::endl;  return GDF_CUDA_ERROR;
-      default:  std::cerr<< "Pagerank failed"<<std::endl;  return GDF_CUDA_ERROR;
+      case -1: CUGRAPH_FAIL("Error : bad parameters in Pagerank");
+      case 1: CUGRAPH_FAIL("Warning : Pagerank did not reached the desired tolerance");
+      default:  CUGRAPH_FAIL("Pagerank exec failed");
     }
 
-  cugraph::detail::copy<WT>(m, d_pr, (WT*)pagerank->data);
+  copy<WT>(m, d_pr, (WT*)pagerank->data);
 
   ALLOC_FREE_TRY(d_val, stream);
 #if 1/* temporary solution till https://github.com/NVlabs/cub/issues/162 is resolved */
@@ -261,10 +259,11 @@ gdf_error gdf_pagerank_impl (gdf_graph *graph,
 #endif
   ALLOC_FREE_TRY(d_leaf_vector, stream);
 
-  return GDF_SUCCESS;
+  
 }
 
-gdf_error gdf_pagerank(gdf_graph *graph, gdf_column *pagerank,
+}
+void pagerank(Graph *graph, gdf_column *pagerank,
         gdf_column *personalization_subset, gdf_column *personalization_values,
         float alpha, float tolerance, int max_iter, bool has_guess) {
   //
@@ -275,12 +274,14 @@ gdf_error gdf_pagerank(gdf_graph *graph, gdf_column *pagerank,
   CUGRAPH_EXPECTS(graph->transposedAdjList != nullptr, "Invalid API parameter");
 
   switch (pagerank->dtype) {
-    case GDF_FLOAT32:   return gdf_pagerank_impl<float>(graph, pagerank,
+    case GDF_FLOAT32:   return detail::pagerank_impl<float>(graph, pagerank,
                                 personalization_subset, personalization_values,
                                 alpha, tolerance, max_iter, has_guess);
-    case GDF_FLOAT64:   return gdf_pagerank_impl<double>(graph, pagerank,
+    case GDF_FLOAT64:   return detail::pagerank_impl<double>(graph, pagerank,
                                 personalization_subset, personalization_values,
                                 alpha, tolerance, max_iter, has_guess);
-    default: return GDF_UNSUPPORTED_DTYPE;
+    default: CUGRAPH_FAIL("Unsupported data type");
   }
 }
+
+} //namespace cugraph