+++ python/cugraph/community/louvain_wrapper.pyx
@@ -16,7 +16,7 @@
 # cython: embedsignature = True
 # cython: language_level = 3
 
-from cugraph.community.c_louvain cimport *
+cimport cugraph.community.c_louvain as c_louvain
 from cugraph.structure.c_graph cimport *
 from cugraph.utilities.column_utils cimport *
 from libcpp cimport bool
@@ -31,24 +31,24 @@ import numpy as np
 
 def louvain(graph_ptr):
     """
-    Call gdf_louvain
+    Call louvain
     """
 
     cdef uintptr_t graph = graph_ptr
-    cdef gdf_graph* g = <gdf_graph*>graph
+    cdef Graph* g = <Graph*>graph
 
-    err = gdf_add_adj_list(g)
-    libcudf.cudf.check_gdf_error(err)
+    add_adj_list(g)
+    
 
-    # we should add get_number_of_vertices() to gdf_graph (and this should be
+    # we should add get_number_of_vertices() to Graph (and this should be
     # used instead of g.adjList.offsets.size - 1)
     num_verts = g.adjList.offsets.size - 1
 
     df = cudf.DataFrame()
     df['vertex'] = cudf.Series(np.zeros(num_verts, dtype=np.int32))
     cdef gdf_column c_index_col = get_gdf_column_view(df['vertex'])
-    err = g.adjList.get_vertex_identifiers(&c_index_col)
-    libcudf.cudf.check_gdf_error(err)
+    g.adjList.get_vertex_identifiers(&c_index_col)
+    
 
     df['partition'] = cudf.Series(np.zeros(num_verts,dtype=np.int32))
     cdef gdf_column c_louvain_parts_col = get_gdf_column_view(df['partition'])
@@ -70,17 +70,17 @@ def louvain(graph_ptr):
     cdef float final_modularity_single_precision = 1.0
     cdef double final_modularity_double_precision = 1.0
     cdef int num_level = 0
-    cdef gdf_error error
+    
 
     if single_precision:
-        err = gdf_louvain(<gdf_graph*>g,
-                          <void*>&final_modularity_single_precision,
-                          <void*>&num_level, &c_louvain_parts_col)
+        c_louvain.louvain(<Graph*>g,
+                  <void*>&final_modularity_single_precision,
+                  <void*>&num_level, &c_louvain_parts_col)
     else:
-        err = gdf_louvain(<gdf_graph*>g,
-                          <void*>&final_modularity_double_precision,
-                          <void*>&num_level, &c_louvain_parts_col)
-    libcudf.cudf.check_gdf_error(err)
+        c_louvain.louvain(<Graph*>g,
+                  <void*>&final_modularity_double_precision,
+                  <void*>&num_level, &c_louvain_parts_col)
+    
 
     if single_precision:
         return df, <double>final_modularity_single_precision