+++ python/cugraph/community/spectral_clustering_wrapper.pyx
@@ -38,17 +38,17 @@ def spectralBalancedCutClustering(graph_ptr,
                                     kmean_tolerance=.00001,
                                     kmean_max_iter=100):
     """
-    Call gdf_balancedCutClustering_nvgraph
+    Call balancedCutClustering_nvgraph
     """
 
     cdef uintptr_t graph = graph_ptr
-    cdef gdf_graph * g = <gdf_graph*> graph
+    cdef Graph * g = <Graph*> graph
 
     # Ensure that the graph has CSR adjacency list
-    err = gdf_add_adj_list(g)
-    libcudf.cudf.check_gdf_error(err)
+    add_adj_list(g)
+    
 
-    # we should add get_number_of_vertices() to gdf_graph (and this should be
+    # we should add get_number_of_vertices() to Graph (and this should be
     # used instead of g.adjList.offsets.size - 1)
     num_verts = g.adjList.offsets.size - 1
 
@@ -60,10 +60,10 @@ def spectralBalancedCutClustering(graph_ptr,
     cdef gdf_column c_cluster_col = get_gdf_column_view(df['cluster'])
 
     # Set the vertex identifiers
-    err = g.adjList.get_vertex_identifiers(&c_identifier_col)
-    libcudf.cudf.check_gdf_error(err)
+    g.adjList.get_vertex_identifiers(&c_identifier_col)
+    
 
-    err = gdf_balancedCutClustering_nvgraph(g,
+    balancedCutClustering_nvgraph(g,
                                             num_clusters,
                                             num_eigen_vects,
                                             evs_tolerance,
@@ -71,7 +71,7 @@ def spectralBalancedCutClustering(graph_ptr,
                                             kmean_tolerance,
                                             kmean_max_iter,
                                             &c_cluster_col)
-    libcudf.cudf.check_gdf_error(err)
+    
 
     return df
 
@@ -83,17 +83,17 @@ def spectralModularityMaximizationClustering(graph_ptr,
                                                kmean_tolerance=.00001,
                                                kmean_max_iter=100):
     """
-    Call gdf_spectralModularityMaximization_nvgraph
+    Call spectralModularityMaximization_nvgraph
     """
 
     cdef uintptr_t graph = graph_ptr
-    cdef gdf_graph * g = <gdf_graph*> graph
+    cdef Graph * g = <Graph*> graph
 
     # Ensure that the graph has CSR adjacency list
-    err = gdf_add_adj_list(g)
-    libcudf.cudf.check_gdf_error(err)
+    add_adj_list(g)
+    
 
-    # we should add get_number_of_vertices() to gdf_graph (and this should be
+    # we should add get_number_of_vertices() to Graph (and this should be
     # used instead of g.adjList.offsets.size - 1)
     num_verts = g.adjList.offsets.size - 1
 
@@ -105,10 +105,10 @@ def spectralModularityMaximizationClustering(graph_ptr,
     cdef gdf_column c_cluster_col = get_gdf_column_view(df['cluster'])
 
     # Set the vertex identifiers
-    err = g.adjList.get_vertex_identifiers(&c_identifier_col)
-    libcudf.cudf.check_gdf_error(err)
+    g.adjList.get_vertex_identifiers(&c_identifier_col)
+    
 
-    err = gdf_spectralModularityMaximization_nvgraph(g,
+    spectralModularityMaximization_nvgraph(g,
                                                      num_clusters,
                                                      num_eigen_vects,
                                                      evs_tolerance,
@@ -116,57 +116,57 @@ def spectralModularityMaximizationClustering(graph_ptr,
                                                      kmean_tolerance,
                                                      kmean_max_iter,
                                                      &c_cluster_col)
-    libcudf.cudf.check_gdf_error(err)
+    
 
     return df
 
 def analyzeClustering_modularity(graph_ptr, n_clusters, clustering):
     """
-    Call gdf_AnalyzeClustering_modularity_nvgraph
+    Call analyzeClustering_modularity_nvgraph
     """
     cdef uintptr_t graph = graph_ptr
-    cdef gdf_graph * g = <gdf_graph*> graph
+    cdef Graph * g = <Graph*> graph
 
     # Ensure that the graph has CSR adjacency list
-    err = gdf_add_adj_list(g)
-    libcudf.cudf.check_gdf_error(err)
+    add_adj_list(g)
+    
 
     cdef gdf_column c_clustering_col = get_gdf_column_view(clustering)
     cdef float score
-    err = gdf_AnalyzeClustering_modularity_nvgraph(g, n_clusters, &c_clustering_col, &score)
-    libcudf.cudf.check_gdf_error(err)
+    analyzeClustering_modularity_nvgraph(g, n_clusters, &c_clustering_col, &score)
+    
     return score
 
 def analyzeClustering_edge_cut(graph_ptr, n_clusters, clustering):
     """
-    Call gdf_AnalyzeClustering_edge_cut_nvgraph
+    Call analyzeClustering_edge_cut_nvgraph
     """
     cdef uintptr_t graph = graph_ptr
-    cdef gdf_graph * g = <gdf_graph*> graph
+    cdef Graph * g = <Graph*> graph
 
     # Ensure that the graph has CSR adjacency list
-    err = gdf_add_adj_list(g)
-    libcudf.cudf.check_gdf_error(err)
+    add_adj_list(g)
+    
 
     cdef gdf_column c_clustering_col = get_gdf_column_view(clustering)
     cdef float score
-    err = gdf_AnalyzeClustering_edge_cut_nvgraph(g, n_clusters, &c_clustering_col, &score)
-    libcudf.cudf.check_gdf_error(err)
+    analyzeClustering_edge_cut_nvgraph(g, n_clusters, &c_clustering_col, &score)
+    
     return score
 
 def analyzeClustering_ratio_cut(graph_ptr, n_clusters, clustering):
     """
-    Call gdf_AnalyzeClustering_ratio_cut_nvgraph
+    Call analyzeClustering_ratio_cut_nvgraph
     """
     cdef uintptr_t graph = graph_ptr
-    cdef gdf_graph * g = <gdf_graph*> graph
+    cdef Graph * g = <Graph*> graph
 
     # Ensure that the graph has CSR adjacency list
-    err = gdf_add_adj_list(g)
-    libcudf.cudf.check_gdf_error(err)
+    add_adj_list(g)
+    
 
     cdef gdf_column c_clustering_col = get_gdf_column_view(clustering)
     cdef float score
-    err = gdf_AnalyzeClustering_ratio_cut_nvgraph(g, n_clusters, &c_clustering_col, &score)
-    libcudf.cudf.check_gdf_error(err)
+    analyzeClustering_ratio_cut_nvgraph(g, n_clusters, &c_clustering_col, &score)
+    
     return score