+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - PR #564 Update cudf type aliases
 - PR #562 Remove pyarrow dependency so we inherit the one cudf uses
 - PR #576 Remove adj list conversion automation from c++
+- PR #587 API upgrade
 - PR #585 Remove BUILD_ABI references from CI scripts
 - PR #591 Adding initial GPU metrics to benchmark utils
 
@@ -16,6 +17,7 @@
 - PR #573 Fix pagerank test and symmetrize for cudf 0.11
 - PR #574 dev env update
 - PR #580 Changed hardcoded test output file to a generated tempfile file name
+- PR #595 Updates to use the new RMM Python reinitialize() API
 
 # cuGraph 0.10.0 (16 Oct 2019)